+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.2.1
 * Add option to override the default background executor, see #292
+* Don't keep a strong reference to finished jobs to allow freeing memory, see #299
 
 ## 1.2.0 (2017-10-05)